Refactor class and add interface
